+++ plugins/nf-validation/src/main/nextflow/validation/FormatValidators/FilePathValidator.groovy
@@ -11,8 +11,8 @@ public class FilePathValidator implements FormatValidator {
 
     @Override
     public Optional<String> validate(final String subject) {
-        if (subject.startsWith('s3://')) {
-            log.debug("S3 paths are not supported by 'FilePathValidator': '${subject}'")
+        if (subject.matches("(s3://|az://|gs://).*")) {
+            log.debug("Cloud storage paths are not supported by 'FilePathValidator': '${subject}'")
             return Optional.empty()
         }
         Path file = Nextflow.file(subject) as Path