+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Version 1.1.4
 
+## Improvements
+
+- No longer does false validation on Azure and GCP cloud storage paths ([#171](https://github.com/nextflow-io/nf-validation/pull/171))
+
 ## Vulnerability fix
 
 - Updated the org.json package to version `20240303` ([#172](https://github.com/nextflow-io/nf-validation/pull/172))